Documentation.Haddock.Doc

Documentation

docParagraph :: DocH mod id -> DocH mod id

docAppend :: DocH mod id -> DocH mod id -> DocH mod id

docConcat :: [DocH mod id] -> DocH mod id

metaDocConcat :: [MetaDoc mod id] -> MetaDoc mod id

Like docConcat but also joins the Meta info.

metaDocAppend :: MetaDoc mod id -> MetaDoc mod id -> MetaDoc mod id

We do something perhaps unexpected here and join the meta info in ‘reverse’: this results in the metadata from the ‘latest’ paragraphs taking precedence.

emptyMetaDoc :: MetaDoc mod id

metaAppend :: Meta -> Meta -> Meta

This is not a monoidal append, it uses <|> for the _version.

metaConcat :: [Meta] -> Meta

Concat using metaAppend.