Documentation.Haddock.Parser

Description

Parser used for Haddock comments. For external users of this library, the most commonly used combination of functions is going to be

toRegular . parseParas

Documentation

parseString :: String -> DocH mod Identifier

Parse a text paragraph. Actually just a wrapper over parseStringBS which drops leading whitespace and encodes the string to UTF8 first.

parseParas

Arguments

:: String
String to parse

-> MetaDoc mod Identifier

Main entry point to the parser. Appends the newline character to the input string.

overIdentifier :: (String -> Maybe a) -> DocH mod Identifier -> DocH mod a

Maps over DocIdentifiers over String with potentially failing conversion using user-supplied function. If the conversion fails, the identifier is deemed to not be valid and is treated as a regular string.

toRegular :: DocH mod Identifier -> DocH mod String

Drops the quotes/backticks around all identifiers, as if they were valid but still Strings.

type Identifier = (Char, String, Char)

Identifier string surrounded with opening and closing quotes/backticks.