Documentation.Haddock.Types

Description

Exposes documentation data types used for (some) of Haddock.

Documentation

newtype Meta

With the advent of Version, we may want to start attaching more meta-data to comments. We make a structure for this ahead of time so we don't have to gut half the core each time we want to add such info.

Constructors

Meta

Fields

_version :: Maybe Version

data MetaDoc mod id

Constructors

MetaDoc

Fields

_meta :: Meta

_doc :: DocH mod id

overDoc :: (DocH a b -> DocH c d) -> MetaDoc a b -> MetaDoc c d

type Version = [Int]

data Hyperlink

Constructors

Hyperlink

Fields

hyperlinkUrl :: String

hyperlinkLabel :: Maybe String

data Picture

Constructors

Picture

Fields

pictureUri :: String

pictureTitle :: Maybe String

data Header id

Constructors

Header

Fields

headerLevel :: Int

headerTitle :: id

data Example

Constructors

Example

Fields

exampleExpression :: String

exampleResult :: [String]

data DocH mod id

Constructors

DocEmpty

DocAppend (DocH mod id) (DocH mod id)

DocString String

DocParagraph (DocH mod id)

DocIdentifier id

DocIdentifierUnchecked mod

DocModule String

DocWarning (DocH mod id)

DocEmphasis (DocH mod id)

DocMonospaced (DocH mod id)

DocBold (DocH mod id)

DocUnorderedList [DocH mod id]

DocOrderedList [DocH mod id]

DocDefList [(DocH mod id, DocH mod id)]

DocCodeBlock (DocH mod id)

DocHyperlink Hyperlink

DocPic Picture

DocAName String

DocProperty String

DocExamples [Example]

DocHeader (Header (DocH mod id))